updating code field to BIGINT to support long numbers

+++ back/src/app/models/Product.js
@@ -7,7 +7,7 @@ class Product extends Model {
         name: Sequelize.STRING,
         brand: Sequelize.STRING,
         price: Sequelize.INTEGER,
-        code: Sequelize.INTEGER,
+        code: Sequelize.BIGINT,
         quantity_per_unity: Sequelize.INTEGER,
         unity: Sequelize.STRING,
       },
